*General changes*

- `CWndTabCtrl` has a way simpler API
- `sqktd::EnumSet` to combine `enum class` and `std::bitset`

*Guild windows*
- Generally improve the code by removing the explicitely redefined methods that had the same definition as the original one
- Better architecture to manage the bindings between controls and powers
- Some files have been merged because Java style 1 file = 1 class hinders the ability to have a grasp of the general architecture of the code, especially as FlyFF does not use many subfolders and project organization is wonky

*Other parts*
- A very few server packet handlers have been simplified.

*New feature*
- The guild master can change their logo.

+++ Source/_Common/TreeInformation.cpp
@@ -33,7 +33,7 @@ BOOL CTreeInformation::GetOpen( void ) const
 	return m_bOpen;
 }
 //-----------------------------------------------------------------------------
-int CTreeInformationManager::m_nSelectedTabNumber = 0;
+size_t CTreeInformationManager::m_nSelectedTabNumber = 0;
 CTreeInformationManager::QuestListGroup CTreeInformationManager::m_eQuestListGroup = CTreeInformationManager::CURRENT_QUEST_LIST;
 //-----------------------------------------------------------------------------
 CTreeInformationManager::CTreeInformationManager( void )
